Revised seasonal code and added concatenation script
Revised the seasonal code in averageBCCA.py to create an extra
temporary cdf, which includes two years of data. This was required to
accurately calculate seasonal averages across a two year span in cases
for Dec-Feb (i.e., winter). The logic for time indexing has also been
revised accordingly and the winter season (DJF) was moved to the end of
the list. The season names were also changed to numeric strings to
avoid filename sorting problems when the concatenation is eventually
run. Added a concatenation script, which will put all of the seasonal,
annual, and monthly cdf files together.