Added new_gmo period arg and patched ncks call
Added a new period called new_gmo to handle the gridded met obs. Also
patched the ncks call to not write blanks in the intermediate and
output netcdf files. This was needed as it was writing zeros to the
output files when it encountered the 1e20 fill data in the source data.